move catch error to su
